light: inform core of current color_mode
While properties allowed the core to be informed of the current color
setting, work_mode was not translated into its corresponding Light
property color_mode.  As a result the icon for a light reconfigured
from a dark red to a warm white would remain dark red, and the UI
variant for changing light status would always start on the color
wheel rather than the color temperature.

Implement the property getter for color_mode to fix this.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>